rename manual & auto tags folders
