+++ src/app/components/home/meta/meta.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, Output, EventEmitter, ChangeDetectorRef, ViewChild, ElementRef } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
-import { ManualTagsService } from '../manual-tags/manual-tags.service';
+import { ManualTagsService } from '../tags-manual/manual-tags.service';
 
 import { StarRating, ImageElement } from '../../common/final-object.interface';
 import { YearEmission } from '../views/details/details.component';
