+++ src/app/components/home/tags-auto/tag-display.pipe.ts
