+++ src/app/components/home/tags-auto/tags.component.ts
@@ -3,9 +3,9 @@ import { Component, EventEmitter, Input, Output, OnDestroy, ViewChild, ElementRe
 import { AutoTagsService, WordAndFreq } from './autotags.service';
 import { AutoTagsSaveService } from './tags-save.service';
 
-import { ImageElement } from '../../../components/common/final-object.interface';
+import { ImageElement } from '../../common/final-object.interface';
 
-import { slowFadeIn, donutAppear } from '../../../components/common/animations';
+import { slowFadeIn, donutAppear } from '../../common/animations';
 
 @Component({
   selector: 'app-tags-component',
