+++ src/app/components/home/tags-manual/view-tags.component.html
